Backend/ProductSceneGenerator/requirements.txt
accelerate==0.34.2
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
diffusers==0.30.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.2
psutil==6.0.0
PyMatting==1.1.12
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
sympy==1.13.3
tifffile==2024.9.20
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
